Adiciona validação de UUID e dependência uuid nas rotas de contato

+++ src/routes/contato.routes.js
@@ -1,7 +1,39 @@
 const express = require('express');
 const prisma = require('../database/prisma');
+const { validate: isUUID, validate } = require('uuid');
+const validateContato = require('../middlewares/validateContato');
+const sanitizeCNPJMiddleware = require('../middlewares/sanitizeCNPJ');
+const { format: formatCNPJ } = require('@fnando/cnpj');
 
 const router = express.Router();
+
+router.post('/', sanitizeCNPJMiddleware, validateContato, async (req, res) => {
+  try {
+    const { nome, email, cnpj } = req.body;
+
+    const novoContato = await prisma.contato.create({
+      data: { nome, email, cnpj },
+    });
+    return res.status(201).json(novoContato);
+  } catch (error) {
+    status = 500;
+    message = 'Erro interno ao criar contato.';
+
+    switch (error?.code) {
+      case 'P2002':
+        status = 409;
+        message = `Usuário já existe`;
+        break;
+    }
+
+    console.error('Erro ao criar contato:', error);
+    return res.status(status).json({ error: message });
+  }
+});
+
+router.put('/:id', async (req, res) => {
+  // Atualização de contato
+});
 
 router.get('/', async (req, res) => {
   try {
@@ -25,7 +57,11 @@ router.get('/', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   const { id } = req.params;
-  console.log(id);
+
+  if (!isUUID(id)) {
+    return res.status(400).json({ error: 'Identificador inválido.' });
+  }
+
   try {
     await prisma.contato.delete({
       where: { id },
